requirementes.txt
asgiref==3.3.1
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
click==7.1.2
colorama==0.4.4
cryptography==3.4.6
CurrencyConverter==0.15
decorator==4.4.2
Django==2.2.1
django-cors-headers==3.7.0
djangorestframework==3.12.2
djangorestframework-jwt==1.11.0
djongo==1.3.4
dynaconf==3.1.2
hvac==0.10.8
idna==2.10
Jinja2==2.11.3
knox==0.1.11
loguru==0.5.3
MarkupSafe==1.1.1
numpy==1.20.1
panda==0.3.1
pandas==1.2.2
pycparser==2.20
PyJWT==1.7.1
pymongo==3.11.3
python-dateutil==2.8.1
pytz==2019.1
requests==2.25.1
shortuuid==1.0.1
six==1.15.0
sqlparse==0.3.0
urllib3==1.26.3
validators==0.18.2
win32-setctime==1.0.3
